requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
cloudinary==1.22.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap4==2.3.1
django-heroku==0.3.1
django-registration==3.1.1
django-tinymce==3.1.0
django-uuidfield==0.5.0
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.11
whitenoise==5.2.0
wrapt==1.12.1
